FancyTerminalEventHandler: manually break lines in progress bar
Manually break lines in the progress bar (and only in the progress bar!)
that are longer than the terminal width. In this way, we avoid any
confusion about when the terminal advances a line; in particular, when
the terminal is wider than we believe, we do not erase back into the
scrollback buffer.

Change-Id: I19f21bdbd6e6af4665d1e712b921f1a15c7d901c