Add interfaces for the build event protocol
Bazel in the will provide a machine-readable stream of important build
events. These interfaces set up the framework and expectations about
the produced events and the entities distributing those events.

Note that at the moment we only request a string representation of the
events to be posted, but once the binary format gets agreed upon, the
interface will be extended to also require a binary representation,
e.g., in the form of a protocol buffer.

Change-Id: If2c3b2e11c31b0136b57eadeef2d2f8f8fe5e2e7