Show the starting of the build in the build event stream
Make the BuildStartingEvent visible in the event stream by making it
an instance of the BuildEvent interface. It is the initial event of the
build event stream hence it also announces that progress updates will
follow. Its regular decedents are the expansion of the target patterns
provided at the request.

Change-Id: I237d8559b71ac82b10fdc492492b8435d6d1483f