Make singlejar aware of FreeBSD
The C implementation of singlejar verifies that it is built for a
supported system; currently those are 64-bit linux and apple systems.
However, the code is portable enough that it can easily be adapted to
run on 64-bit FreeBSD systems as well. In fact, only the path of a
single header file is different, and we need to link with -lm.

Fixes #3096.

Change-Id: Id69542ad940f1161aceb55eb8f5dabda2f486747